functions/lro_status_checker/requirements.txt
google-cloud-storage==2.9.0
google-cloud-documentai==2.15.1
google-api-core==2.11.0